# services/price-service/docs/swagger.yaml
basePath: /internal
definitions:
  handlers.BasketItem:
    properties:
      itemId:
        type: string
      name:
        type: string
      quantity:
        minimum: 1
        type: integer
    required:
    - itemId
    - name
    - quantity
    type: object
...
  handlers.ItemPriceInfo:
    properties:
      basePrice:
        type: integer
      discountPrice:
        type: integer
      effectivePrice:
        type: integer
      hasDiscount:
        type: boolean
      itemId:
        type: string
      itemName:
        type: string
      lineTotal:
        type: integer
      quantity:
        type: integer
    type: object
...
  handlers.Location:
    properties:
      latitude:
        maximum: 90
        minimum: -90
        type: number
      longitude:
        maximum: 180
        minimum: -180
        type: number
    required:
    - latitude
    - longitude
    type: object
  handlers.MissingItem:
    properties:
      isOptional:
        type: boolean
      itemId:
        type: string
      itemName:
        type: string
      penalty:
        type: integer
    type: object
  handlers.MultiStoreResult:
    properties:
      algorithmUsed:
        type: string
      combinedTotal:
        type: integer
      coverageRatio:
        type: number
      stores:
        items:
          $ref: '#/definitions/handlers.StoreAllocation'
        type: array
      unassignedItems:
        items:
          $ref: '#/definitions/handlers.MissingItem'
        type: array
    type: object
  handlers.OptimizeRequest:
    properties:
      basketItems:
        items:
          $ref: '#/definitions/handlers.BasketItem'
        maxItems: 100
        minItems: 1
        type: array
      chainSlug:
        type: string
      location:
        $ref: '#/definitions/handlers.Location'
      maxDistance:
        type: number
      maxStores:
        type: integer
    required:
    - basketItems
    - chainSlug
    type: object
...
  handlers.StoreAllocation:
    properties:
      distance:
        type: number
      items:
        items:
          $ref: '#/definitions/handlers.ItemPriceInfo'
        type: array
      storeId:
        type: string
      storeTotal:
        type: integer
      visitOrder:
        type: integer
    type: object
...
info:
  contact: {}
  description: Internal API for price data management, ingestion monitoring, and basket
    optimization.
  title: Price Service API
  version: "1.0"
paths:
...
  /internal/basket/cache/refresh/{chainSlug}:
    post:
      consumes:
      - application/json
      description: Refreshes the price cache for a specific chain
      parameters:
      - description: Chain slug identifier
        in: path
        name: chainSlug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cache refreshed
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Cache not initialized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh chain cache
      tags:
      - cache
  /internal/basket/cache/warmup:
    post:
      consumes:
      - application/json
      description: Triggers a full cache warmup for all chains
      produces:
      - application/json
      responses:
        "200":
          description: Cache warmed up
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Cache not initialized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Warm up price cache
      tags:
      - cache
  /internal/basket/optimize/multi:
    post:
      consumes:
      - application/json
      description: Finds the optimal distribution of basket items across multiple
        stores
      parameters:
      - description: Optimization request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.OptimizeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MultiStoreResult'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Cache unavailable
          schema:
            additionalProperties:
              type: string
            type: object
        "504":
          description: Optimization timed out
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Optimize basket across multiple stores
      tags:
      - basket
  /internal/basket/optimize/single:
    post:
      consumes:
      - application/json
      description: Finds the best single store for a basket of items based on price
        and coverage
      parameters:
      - description: Optimization request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.OptimizeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Optimization results
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Cache unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Optimize basket for single store
      tags:
      - basket
...
swagger: "2.0"